Changed offset from double to long long. That should be plenty for tracking the overflows of 32 bit counter.
git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1950 a8e31b63-e6ca-409b-a01b-2eec1b506579

+++ src/Profile/MultipleCounters.cpp
@@ -893,7 +893,7 @@ void MultipleCounterLayer::papiMCL(int tid, double values[]){
 
 void MultipleCounterLayer::papiWallClockMCL(int tid, double values[]){
 static long long oldvalue = 0L;
-static double offset = 0;
+static long long offset = 0;
 long long newvalue = 0L;
 #ifdef TAU_PAPI
   newvalue = PAPI_get_real_usec();
