+++ src/Profile/PapiLayer.cpp
@@ -215,7 +215,7 @@ long long PapiLayer::getWallClockTime(void)
 { // Returns the wall clock time from PAPI interface
   static int initflag = PapiLayerInit();
   static long long oldvalue = 0L;
-  static double offset = 0;
+  static long long offset = 0;
   long long newvalue = 0L;
 #ifdef TAU_PAPI
   newvalue = PAPI_get_real_usec();
